Add HLS alternate media parsing support.
Allow parsing of the #EXT-X-MEDIA directive in HLS playlists for all
types. The alternate medias are stored in the master playlist and for
each varianta list with its associated media is created.

Also fix the parsing of the uri attribute which is optional and not
mandatory. It was breaking the playing of the advanced HLS playlist in
the demo samples list.