Fixed the Mp4 embedded WebVTT playback
There was a nasty error which I found as a side effect of fixing issue #1185. One of the
last when Mp4/WebVTT was implemented was to merge the "mp4webvtt" and "webvtt" packages,
which resulted in the TextTrackRenderer not finding the Subtitle parser class.

On top of that, the DashChunkSource did not have a branch for the MP4 variant of webvtt
(only ttml).

Even though Smooth Streaming uses TTML, I am wondering if the WebVTT variant could be
found in the wild. I was not able to find any fourCC related to WebVTT.

Issue: #1185
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112957377