Support multiple BaseURL elements in MPD
All BaseURL elements defined in the MPD is now stored as a List from which a BaseUrlSelector is used to select one of the available base URLs on requests.

URLs at each level of the MPD are resolved with respect to the BaseURL elements specified at that level. Relative BaseURL elements are themselves resolved using the BaseURL elements of the level above. If no BaseURL elements are defined at a certain level, BaseURL elements from the level above are inherited.