Add loading error reports to ExtractorSampleSource
This CL also implements loading errors in the DemoApp by logging them.
The sourceId in the ExtractorSampleSource interface allows a class to
implement the same signature for all the SampleSources. The
eventSourceId in the ExtractorSampleSource is unlikely to be useful,
considering it is expected to use one single source. However, it was
added for two reasons:

* Consistency, since the ChunkSampleSource
* In the strange case that two ExtractorSampleSources live at the same
  time, it would be possible to diferentiate them.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116366928