Consider udta box as leaf
The MP4 standard considers the udta box as a regular container box.
Quicktime, however, considers it as a container (of only leaf atoms)
that can have a terminating 32 bit integer with value 0. Since this
breaks the principle of not having content in container boxes, this
CL considers the udta box as a leaf box that contains other boxes and
does the parsing manually.

Issue: #1315
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116532740