Access the buffer size only if it's a new buffer.
Pre-API 21, AudioTrack.handleBuffer reads the whole input buffer into a
temporary buffer before writing to AudioTrack. This operation advances the
position on the ByteBuffer passed to handleBuffer to its limit.
LibflacAudioTrackRenderer passes the remaining number of bytes in the buffer as
the buffer size. This caused a buffer to be incorrectly treated as consumed on
the second time it was passed to handleBuffer after having been only partially
consumed.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=127291150