Fix unnecessary boolean expression
This change fixes boolean expressions like `a && a` and `a || a` where both
arguments are the same, by replacing the entire expression with `a`.

The clean up is being done in preparation for enabling a build error.

*Note to reviewers* the intent may have been to use two different expressions
(e.g. `a && b`). If this isn't the best fix, please suggest an edit in
critique.

More information:
https://docs.google.com[]document/d/1FN2oqU8QmJ2aVSx969rORp-nWcfgdVjqZ6KOFyxYcHY/edit?usp=sharing

Tested:
    TAP train for global presubmit queue
    []    All failing tests were failing before this change

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=127787709