Largely mitigate seeking-in-live edge case in v1.
When a client implements seeking within the live window in v1
with startAtLiveEdge=true, the following edge case can occur:

1. First audio chunk is fully downloaded. startAtLiveEdge is
   toggled to false.
2. First video chunk is still downloading. startAtLiveEdge is
   still true.
3. User initiates a seek. The video DashChunkSource will seek
   to the live edge, whilst the audio DashChunkSource will
   seek to the specified position, and hence they'll end up
   out of sync.

A complete fix seems non-trivial and we're fixing this properly
in v2. This change resolves the issue for all cases except when
the user initiated seek is to t=0. Note that for continuously
running live streams (e.g. live TV) t=0 is typically not part
of the live window, and so the issue is fully resolved for such
streams.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128192189